divide reducer to small reducers
