+++ src/reducer/game/game.js
@@ -1,15 +1,11 @@
 const initialState = {
-  isAuthorizationRequired: false,
   mistakes: 0,
-  questions: [],
   step: -1,
 };
 
 const ACTION_TYPE = new Map([
   [`incrementMistake`, `INCREMENT_MISTAKES`],
   [`incrementStep`, `INCREMENT_STEP`],
-  [`loadQuestions`, `LOAD_QUESTIONS`],
-  [`requiredAuthorization`, `REQUIRED_AUTHORIZATION`],
   [`reset`, `RESET`],
 ]);
 
@@ -37,16 +33,6 @@ const ActionCreator = {
     };
   },
 
-  loadQuestions: (questions) => ({
-    type: ACTION_TYPE.get(`loadQuestions`),
-    payload: questions,
-  }),
-
-  requiredAuthorization: (status) => ({
-    type: ACTION_TYPE.get(`requiredAuthorization`),
-    payload: status,
-  }),
-
   restart: () => ({
     type: ACTION_TYPE.get(`reset`),
   }),
@@ -62,14 +48,6 @@ const isGenreAnswerCorrect = (userAnswer, question) => {
   ));
 };
 
-const Operation = {
-  loadQuestions: () => (dispatch, _getState, api) =>
-    api.get(`/questions`)
-      .then((response) => {
-        dispatch(ActionCreator.loadQuestions(response.data));
-      }),
-};
-
 const reducer = (state = initialState, action) => {
   switch (action.type) {
     case ACTION_TYPE.get(`incrementStep`):
@@ -82,16 +60,6 @@ const reducer = (state = initialState, action) => {
         mistakes: state.mistakes + action.payload,
       });
 
-    case ACTION_TYPE.get(`loadQuestions`):
-      return Object.assign({}, state, {
-        questions: action.payload
-      });
-
-    case ACTION_TYPE.get(`requiredAuthorization`):
-      return Object.assign({}, state, {
-        isAuthorizationRequired: action.payload,
-      });
-
     case ACTION_TYPE.get(`reset`):
       return Object.assign({}, initialState);
   }
@@ -102,6 +70,5 @@ export {
   ActionCreator,
   isArtistAnswerCorrect,
   isGenreAnswerCorrect,
-  Operation,
   reducer,
 };
